MyProject.Test/Application.Test/Services.Test/PermissionServiceTests.cs
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using Moq;
using MyProject.Application.Services;
using MyProject.Domain.Entities;
using MyProject.Domain.Interfaces;
using Xunit;

namespace MyProject.Tests.Services
{
    public class PermissionServiceTests
    {
        [Theory]
        [InlineData(1)]
        [InlineData(2)]
        public async Task GetByIdAsync_ExistingId_ReturnsPermission(int permissionId)
        {
            // Arrange
            var mockRepository = new Mock<IPermissionRepository>();
            var service = new PermissionService(mockRepository.Object);

            var expectedPermission = new Permission { Id = permissionId, Name = "TestPermission" };
            mockRepository.Setup(repo => repo.GetByIdAsync(permissionId)).ReturnsAsync(expectedPermission);

            // Act
            var result = await service.GetByIdAsync(permissionId);

            // Assert
            Assert.NotNull(result);
            Assert.Equal(permissionId, result.Id);
            // Add more specific assertions based on your expected behavior
        }

        [Fact]
        public async Task GetAllAsync_ReturnsAllPermissions()
        {
            // Arrange
            var mockRepository = new Mock<IPermissionRepository>();
            var service = new PermissionService(mockRepository.Object);

            var expectedPermissions = new List<Permission>
            {
                new Permission { Id = 1, Name = "Permission 1" },
                new Permission { Id = 2, Name = "Permission 2" },
                // Add more sample permissions as needed
            };
            mockRepository.Setup(repo => repo.GetAllAsync()).ReturnsAsync(expectedPermissions);

            // Act
            var result = await service.GetAllAsync();

            // Assert
            Assert.NotNull(result);
            // Add more specific assertions based on your expected behavior
        }

        [Fact]
        public async Task CreateAsync_ValidPermission_ReturnsPermissionId()
        {
            // Arrange
            var mockRepository = new Mock<IPermissionRepository>();
            var service = new PermissionService(mockRepository.Object);

            var newPermission = new Permission { Name = "New Permission" }; // Replace with test data
            var createdPermissionId = 1; // Replace with an expected ID for testing purposes
            mockRepository.Setup(repo => repo.AddAsync(newPermission)).ReturnsAsync(new Permission { Id = createdPermissionId });

            // Act
            var result = await service.CreateAsync(newPermission);

            // Assert
            Assert.Equal(createdPermissionId, result);
            // Add more specific assertions based on your expected behavior
        }

        // Write similar tests for UpdateAsync and DeleteAsync methods

        // Remember to add specific assertions to validate the responses and behaviors of each method.
    }
}
